Refactor: Update UserSession balance type to float
This commit changes the data type for user balance from `double` to `float` in `DepositFundsController.java`.

**Specific Changes:**

*   **`app/src/main/java/com/example/mandelamoney/controller/DepositFundsController.java`:**
    *   In the `Executors.newSingleThreadExecutor().execute(() -> { ... })` block, the local variable `updated` which stores the result of `UserSession.updateBalance(context)` has been changed from `double` to `float`.
    *   The `setUserBalance` method on the `User` object `u` is now called with this `float` value.

+++ app/src/main/java/com/example/mandelamoney/controller/DepositFundsController.java
@@ -283,7 +283,7 @@ public class DepositFundsController {
 
                 // 2) App: refresh session balance so UI shows the new total
                 Executors.newSingleThreadExecutor().execute(() -> {
-                    double updated = UserSession.updateBalance(context);
+                    float updated = UserSession.updateBalance(context);
                     User u = UserSession.getUser();
                     if (u != null) u.setUserBalance(updated);
                     Log.d(TAG, "Session balance refreshed to: " + updated);
